Lots of changes, main changes: Addition of RestartFrameWinBuild.java & TitleFrame1WinBuild.java for ease of styling. Both were made using winBuilder in addition to adding the code from the original files. Original files, RestartFrame.java & TitleFrame1.java were left but all code was commented out to ensure no missed references.
Now main method creates a highScores.txt file wherever it is (Same when it is exported)
that it can write and read too. Also styles the text going into the file for ease of reading.

WARNING!! When fully compiled and exported recieved warnings about missing classes,
the original ones that were commented out. Still trying to figure out if it broke anything
but it seems to all still work fine. Couldn't find any other errors except when exporting.